# src/main/resources/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Products & Services API",
    "description": "REST API for managing products, services, classifications, and value components",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8888",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/v2/products-services/value-movement/list": {
      "post": {
        "tags": [
          "value-movement-controller"
        ],
        "operationId": "findAllPaginated",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValueMovementDataRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/products-services/value-component/list": {
      "post": {
        "tags": [
          "value-component-controller"
        ],
        "operationId": "findAllPaginated_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValueComponentDataRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/products-services/value-component-type/list": {
      "post": {
        "tags": [
          "value-component-type-controller"
        ],
        "operationId": "findAllPaginated_2",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValueComponentTypeDataRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/products-services/value-application/list": {
      "post": {
        "tags": [
          "value-application-controller"
        ],
        "operationId": "findAllPaginated_3",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValueApplicationDataRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/products-services/service/list": {
      "post": {
        "tags": [
          "service-controller"
        ],
        "operationId": "findAllPaginated_4",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceDataRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/products-services/service-classification/list": {
      "post": {
        "tags": [
          "service-classification-controller"
        ],
        "operationId": "findAllPaginated_5",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceClassificationDataRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/products-services/product/list": {
      "post": {
        "tags": [
          "product-controller"
        ],
        "operationId": "findAllPaginated_6",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductDataRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/products-services/product-type/list": {
      "post": {
        "tags": [
          "product-type-controller"
        ],
        "operationId": "findAllPaginated_7",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductTypeDataRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/products-services/product-component/list": {
      "post": {
        "tags": [
          "product-component-controller"
        ],
        "operationId": "findAllPaginated_8",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductComponentDataRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/products-services/product-classification/list": {
      "post": {
        "tags": [
          "product-classification-controller"
        ],
        "operationId": "findAllPaginated_9",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductClassificationDataRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/products-services/product-and-service-classification/list": {
      "post": {
        "tags": [
          "product-and-service-classification-controller"
        ],
        "operationId": "findAllPaginated_10",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductAndServiceClassificationDataRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/products-services/currency/list": {
      "post": {
        "tags": [
          "currency-controller"
        ],
        "operationId": "findAllPaginated_11",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CurrencyDataRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/products-services/currency-classification/list": {
      "post": {
        "tags": [
          "currency-classification-controller"
        ],
        "operationId": "findAllPaginated_12",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CurrencyClassificationDataRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/products-services/country/list": {
      "post": {
        "tags": [
          "country-controller"
        ],
        "operationId": "findAllPaginated_13",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CountryDataRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/products-services/business-entity-category/list": {
      "post": {
        "tags": [
          "business-entity-category-controller"
        ],
        "operationId": "findAllPaginated_14",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BusinessEntityCategoryDataRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/products-services/value-component/{id}/status": {
      "patch": {
        "tags": [
          "value-component-controller"
        ],
        "operationId": "updateStatus",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/products-services/service-classification/{id}/status": {
      "patch": {
        "tags": [
          "service-classification-controller"
        ],
        "operationId": "updateStatus_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/products-services/product-classification/{id}/status": {
      "patch": {
        "tags": [
          "product-classification-controller"
        ],
        "operationId": "updateStatus_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/products-services/currency/{id}/status": {
      "patch": {
        "tags": [
          "currency-controller"
        ],
        "operationId": "updateStatus_3",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/products-services/value-movement/{id}": {
      "get": {
        "tags": [
          "value-movement-controller"
        ],
        "operationId": "findById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/products-services/value-component/{id}": {
      "get": {
        "tags": [
          "value-component-controller"
        ],
        "operationId": "findById_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/products-services/value-component-type/{id}": {
      "get": {
        "tags": [
          "value-component-type-controller"
        ],
        "operationId": "findById_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/products-services/value-application/{id}": {
      "get": {
        "tags": [
          "value-application-controller"
        ],
        "operationId": "findById_3",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/products-services/service/{id}": {
      "get": {
        "tags": [
          "service-controller"
        ],
        "operationId": "findById_4",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/products-services/service-classification/{id}": {
      "get": {
        "tags": [
          "service-classification-controller"
        ],
        "operationId": "findById_5",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/products-services/product/{id}": {
      "get": {
        "tags": [
          "product-controller"
        ],
        "operationId": "findById_6",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/products-services/product-type/{id}": {
      "get": {
        "tags": [
          "product-type-controller"
        ],
        "operationId": "findById_7",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/products-services/product-component/{id}": {
      "get": {
        "tags": [
          "product-component-controller"
        ],
        "operationId": "findById_8",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/products-services/product-classification/{id}": {
      "get": {
        "tags": [
          "product-classification-controller"
        ],
        "operationId": "findById_9",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/products-services/product-and-service-classification/{id}": {
      "get": {
        "tags": [
          "product-and-service-classification-controller"
        ],
        "operationId": "findById_10",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/products-services/currency/{id}": {
      "get": {
        "tags": [
          "currency-controller"
        ],
        "operationId": "findById_11",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/products-services/currency-classification/{id}": {
      "get": {
        "tags": [
          "currency-classification-controller"
        ],
        "operationId": "findById_12",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/products-services/country/{id}": {
      "get": {
        "tags": [
          "country-controller"
        ],
        "operationId": "findById_13",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/products-services/business-entity-category/{id}": {
      "get": {
        "tags": [
          "business-entity-category-controller"
        ],
        "operationId": "findById_14",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GlobalResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "data": {
            "type": "object"
          },
          "message": {
            "type": "string"
          },
          "errorDetail": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        }
      },
...
      "ValueMovementDataRequest": {
        "type": "object",
        "properties": {
          "pageNo": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "pageSize": {
            "type": "integer",
            "format": "int32",
            "maximum": 500,
            "minimum": 5
          },
          "sortBy": {
            "type": "string"
          },
          "direction": {
            "type": "string"
          },
          "searchText": {
            "type": "string"
          }
        },
        "required": [
          "pageNo",
          "pageSize"
        ]
      },
      "ValueComponentDataRequest": {
        "type": "object",
        "properties": {
          "pageNo": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "pageSize": {
            "type": "integer",
            "format": "int32",
            "maximum": 500,
            "minimum": 5
          },
          "sortBy": {
            "type": "string"
          },
          "direction": {
            "type": "string"
          },
          "searchText": {
            "type": "string"
          }
        },
        "required": [
          "pageNo",
          "pageSize"
        ]
      },
      "ValueComponentTypeDataRequest": {
        "type": "object",
        "properties": {
          "pageNo": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "pageSize": {
            "type": "integer",
            "format": "int32",
            "maximum": 500,
            "minimum": 5
          },
          "sortBy": {
            "type": "string"
          },
          "direction": {
            "type": "string"
          },
          "searchText": {
            "type": "string"
          }
        },
        "required": [
          "pageNo",
          "pageSize"
        ]
      },
      "ValueApplicationDataRequest": {
        "type": "object",
        "properties": {
          "pageNo": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "pageSize": {
            "type": "integer",
            "format": "int32",
            "maximum": 500,
            "minimum": 5
          },
          "sortBy": {
            "type": "string"
          },
          "direction": {
            "type": "string"
          },
          "searchText": {
            "type": "string"
          }
        },
        "required": [
          "pageNo",
          "pageSize"
        ]
      },
      "ServiceDataRequest": {
        "type": "object",
        "properties": {
          "pageNo": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "pageSize": {
            "type": "integer",
            "format": "int32",
            "maximum": 500,
            "minimum": 5
          },
          "sortBy": {
            "type": "string"
          },
          "direction": {
            "type": "string"
          },
          "searchText": {
            "type": "string"
          }
        },
        "required": [
          "pageNo",
          "pageSize"
        ]
      },
      "ServiceClassificationDataRequest": {
        "type": "object",
        "properties": {
          "pageNo": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "pageSize": {
            "type": "integer",
            "format": "int32",
            "maximum": 500,
            "minimum": 5
          },
          "sortBy": {
            "type": "string"
          },
          "direction": {
            "type": "string"
          },
          "searchText": {
            "type": "string"
          }
        },
        "required": [
          "pageNo",
          "pageSize"
        ]
      },
      "ProductDataRequest": {
        "type": "object",
        "properties": {
          "pageNo": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "pageSize": {
            "type": "integer",
            "format": "int32",
            "maximum": 500,
            "minimum": 5
          },
          "sortBy": {
            "type": "string"
          },
          "direction": {
            "type": "string"
          },
          "searchText": {
            "type": "string"
          }
        },
        "required": [
          "pageNo",
          "pageSize"
        ]
      },
      "ProductTypeDataRequest": {
        "type": "object",
        "properties": {
          "pageNo": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "pageSize": {
            "type": "integer",
            "format": "int32",
            "maximum": 500,
            "minimum": 5
          },
          "sortBy": {
            "type": "string"
          },
          "direction": {
            "type": "string"
          },
          "searchText": {
            "type": "string"
          }
        },
        "required": [
          "pageNo",
          "pageSize"
        ]
      },
      "ProductComponentDataRequest": {
        "type": "object",
        "properties": {
          "pageNo": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "pageSize": {
            "type": "integer",
            "format": "int32",
            "maximum": 500,
            "minimum": 5
          },
          "sortBy": {
            "type": "string"
          },
          "direction": {
            "type": "string"
          },
          "searchText": {
            "type": "string"
          }
        },
        "required": [
          "pageNo",
          "pageSize"
        ]
      },
...
      "StatusUpdateRequest": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE",
              "DELETED",
              "REJECTED"
            ]
          }
        }
      }
    }
  }
}